+++ checkout-ui-custom/src/utils/const.js
@@ -6,14 +6,20 @@ const STEPS = {
   PAYMENT: '#/payment'
 };
 
-// OrderForm timeout
-const ORDERFORM_TIMEOUT = 500;
+// TIMEOUT
+const TIMEOUT_500 = 500;
+const TIMEOUT_750 = 750;
 
-// OrderForm configuration app-name
-const CUSTOM_FIELDS_APP = 'customshippinginfo';
+const RICA_APP = 'ricafields';
+
+// Furniture fees Url
+const FURNITURE_FEES = 'http://image.tfgmedia.co.za/image/1/process/500x790?source=http://cdn.tfgmedia.co.za'
+  + '/15/Marketing/HTMLPages/Furniture_Delivery_Fees_tab_image.jpg';
 
 export {
   STEPS,
-  ORDERFORM_TIMEOUT,
-  CUSTOM_FIELDS_APP
+  TIMEOUT_500,
+  TIMEOUT_750,
+  RICA_APP,
+  FURNITURE_FEES
 };