+++ CHANGELOG.md
@@ -20,3 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Functionality associated with RICA
 - Functionality associated with Furniture products
 - Functionality associated with Mixed products
+### Fixed
+
+- Fixed issues after test all the cases
+