+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.1.0] - 2022-01-18
+## [0.1.0] - 2022-XX-XX
 
 ### Changed
 
@@ -16,3 +16,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Initial version of the project created.
+- Functionality associated with TV licence
+- Functionality associated with RICA
+- Functionality associated with Furniture products
+- Functionality associated with Mixed products
+### Fixed
+
+- Fixed issues after test all the cases
+